Separate Product impl
