+++ SeQuant/core/expressions/expr.cpp
@@ -26,8 +26,6 @@
 #include <range/v3/view/reverse.hpp>
 #include <range/v3/view/transform.hpp>
 
-#include <vector>
-
 namespace sequant {
 
 ExprIterator Expr::begin() { return begin_subexpr(); }
@@ -78,209 +76,6 @@ const ExprPtr &Expr::back() const { return at(size() - 1); }
 
 std::wstring Expr::to_latex() const {
   throw Exception("to_latex not implemented for " + type_name());
-}
-
-bool Product::is_commutative() const {
-  bool result = true;
-  const auto nfactors = size();
-  for (size_t f = 0; f != nfactors; ++f) {
-    for (size_t s = 1; result && s != nfactors; ++s) {
-      result &= factors_[f]->commutes_with(*factors_[s]);
-    }
-  }
-  return result;
-}
-
-ExprPtr Product::canonicalize_impl(CanonicalizeOptions opts) {
-  // recursively canonicalize non-tensor subfactors (tensors will be
-  // canonicalized as part of the TN built of all tensor factors of this) ...
-  ranges::for_each(factors_, [this, opts](auto &factor) {
-    if (factor.template is<AbstractTensor>()) {
-      return;
-    }
-    auto bp = factor->canonicalize(opts);
-    if (bp) {
-      SEQUANT_ASSERT(bp->template is<Constant>());
-      this->scalar_ *= std::static_pointer_cast<Constant>(bp)->value();
-    }
-  });
-
-  if (Logger::instance().canonicalize) {
-    std::wcout << "Product canonicalization(" << to_wstring(opts.method)
-               << ") input: " << to_latex() << std::endl;
-  }
-
-  // pull out all scalar factors to the front
-  auto is_scalar = [](const auto &factor) { return factor->is_scalar(); };
-  auto scalars =
-      factors_ | ranges::views::filter(is_scalar) | ranges::to_vector;
-  // scalars commute, so we can reorder them freely
-  ranges::sort(scalars, [](const auto &first, const auto &second) {
-    return *first < *second;
-  });
-
-  factors_ = factors_ | ranges::views::filter([&is_scalar](const auto &factor) {
-               return !is_scalar(factor);
-             }) |
-             ranges::to<decltype(factors_)>;
-
-  // if there are no factors, insert scalars back and return
-  if (factors_.empty()) {
-    factors_.insert(factors_.begin(), scalars.begin(), scalars.end());
-    return {};
-  }
-
-  auto contains_nontensors = ranges::any_of(factors_, [](const auto &factor) {
-    return std::dynamic_pointer_cast<AbstractTensor>(factor) == nullptr;
-  });
-  if (!contains_nontensors) {  // tensor network canonization is a special case
-                               // that's done in
-                               // TensorNetwork
-    auto make_canonical_tn = [this, &opts](auto *tn_null_ptr) {
-      using TN = std::decay_t<std::remove_pointer_t<decltype(tn_null_ptr)>>;
-      ExprPtr canon_factor;
-      TN tn(this->factors_);
-      if constexpr (TN::version() == 3) {
-        canon_factor = tn.canonicalize(
-            TensorCanonicalizer::cardinal_tensor_labels(), opts);
-      } else {
-        using NamedIndexSet = tensor_network::NamedIndexSet;
-        std::shared_ptr<NamedIndexSet> named_indices =
-            !opts.named_indices
-                ? nullptr
-                : std::make_shared<NamedIndexSet>(opts.named_indices->begin(),
-                                                  opts.named_indices->end());
-        canon_factor = tn.canonicalize(
-            TensorCanonicalizer::cardinal_tensor_labels(),
-            opts.method == CanonicalizationMethod::Rapid, named_indices.get());
-      }
-      return std::pair{std::move(tn), canon_factor};
-    };
-    using TN = TensorNetwork;
-    auto [tn, canon_factor] = make_canonical_tn(static_cast<TN *>(nullptr));
-
-    const auto &tensors = tn.tensors();
-    using std::size;
-    SEQUANT_ASSERT(size(tensors) == size(factors_));
-    using std::begin;
-    using std::end;
-    std::transform(begin(tensors), end(tensors), begin(factors_),
-                   [](const auto &tptr) {
-                     auto exprptr = std::dynamic_pointer_cast<Expr>(tptr);
-                     SEQUANT_ASSERT(exprptr);
-                     return exprptr;
-                   });
-    if (canon_factor) scalar_ *= canon_factor->template as<Constant>().value();
-    this->reset_hash_value();
-  } else {  // if contains non-tensors, do commutation-checking resort
-
-    // comparer that respects cardinal tensor labels
-    auto &cardinal_tensor_labels =
-        TensorCanonicalizer::cardinal_tensor_labels();
-    auto local_compare = [&cardinal_tensor_labels](const ExprPtr &first,
-                                                   const ExprPtr &second) {
-      if (first->is<Labeled>() && second->is<Labeled>()) {
-        const auto first_label = first->as<Labeled>().label();
-        const auto second_label = second->as<Labeled>().label();
-        if (first_label == second_label) return *first < *second;
-        const auto first_is_cardinal_it = ranges::find_if(
-            cardinal_tensor_labels,
-            [&first_label](const std::wstring &l) { return l == first_label; });
-        const auto first_is_cardinal =
-            first_is_cardinal_it != ranges::end(cardinal_tensor_labels);
-        const auto second_is_cardinal_it = ranges::find_if(
-            cardinal_tensor_labels, [&second_label](const std::wstring &l) {
-              return l == second_label;
-            });
-        const auto second_is_cardinal =
-            second_is_cardinal_it != ranges::end(cardinal_tensor_labels);
-        if (first_is_cardinal && second_is_cardinal)
-          return first_is_cardinal_it < second_is_cardinal_it;
-        else if (first_is_cardinal && !second_is_cardinal)
-          return true;
-        else if (!first_is_cardinal && second_is_cardinal)
-          return false;
-        else {
-          SEQUANT_ASSERT(!first_is_cardinal && !second_is_cardinal);
-          return *first < *second;
-        }
-      } else
-        return *first < *second;
-    };
-
-    // ... then resort, respecting commutativity
-    using std::begin;
-    using std::end;
-    if (static_commutativity()) {
-      if (is_commutative()) {
-        std::stable_sort(begin(factors_), end(factors_), local_compare);
-      }
-    } else {
-      // must do bubble sort if not commuting to avoid swapping elements across
-      // a noncommuting element
-      bubble_sort(
-          begin(factors_), end(factors_),
-          [&local_compare](const ExprPtr &first, const ExprPtr &second) {
-            bool result = (first->commutes_with(*second))
-                              ? local_compare(first, second)
-                              : false;
-            return result;
-          });
-    }
-  }
-  // reinsert scalar factors at the front
-  factors_.insert(factors_.begin(), scalars.begin(), scalars.end());
-
-  // TODO evaluate product of Tensors (turn this into Products of Products)
-
-  if (Logger::instance().canonicalize)
-    std::wcout << "Product canonicalization(" << to_wstring(opts.method)
-               << ") result: " << to_latex() << std::endl;
-
-  return {};  // side effects are absorbed into the scalar_
-}
-
-void Product::adjoint() {
-  SEQUANT_ASSERT(static_commutativity() == false);  // assert no slicing
-  auto adj_scalar = conj(scalar());
-  using namespace ranges;
-  auto adj_factors =
-      factors() | views::reverse |
-      views::transform([](auto &expr) { return ::sequant::adjoint(expr); });
-  using std::swap;
-  *this =
-      Product(adj_scalar, ranges::begin(adj_factors), ranges::end(adj_factors));
-}
-
-ExprPtr Product::canonicalize(CanonicalizeOptions opt) {
-  return this->canonicalize_impl(opt);
-}
-
-ExprPtr Product::rapid_canonicalize(CanonicalizeOptions opt) {
-  SEQUANT_ASSERT(opt.method == CanonicalizationMethod::Rapid);
-  return this->canonicalize_impl(opt);
-}
-
-void CProduct::adjoint() {
-  auto adj_scalar = conj(scalar());
-  using namespace ranges;
-  // no need to reverse for commutative product
-  auto adj_factors = factors() | views::transform([](auto &&expr) {
-                       return ::sequant::adjoint(expr);
-                     });
-  *this = CProduct(adj_scalar, ranges::begin(adj_factors),
-                   ranges::end(adj_factors));
-}
-
-void NCProduct::adjoint() {
-  auto adj_scalar = conj(scalar());
-  using namespace ranges;
-  // no need to reverse for commutative product
-  auto adj_factors =
-      factors() | views::reverse |
-      views::transform([](auto &&expr) { return ::sequant::adjoint(expr); });
-  *this = NCProduct(adj_scalar, ranges::begin(adj_factors),
-                    ranges::end(adj_factors));
 }
 
 void Sum::adjoint() {
