+++ SeQuant/core/expressions/product.hpp
@@ -35,6 +35,7 @@ class Product : public Expr {
   enum class Flatten { Once, Recursively, Yes = Recursively, No };
 
   using scalar_type = Constant::scalar_type;
+  using factors_type = container::svector<ExprPtr, 2>;
 
   Product() = default;
   virtual ~Product() = default;
@@ -46,12 +47,7 @@ class Product : public Expr {
   /// construct a Product out of zero or more factors (multiplied by 1)
   /// @param factors the factors
   /// @param flatten_tag if Flatten::Yes, flatten the factors
-  Product(ExprPtrList factors, Flatten flatten_tag = Flatten::Yes) {
-    using std::begin;
-    using std::end;
-    for (auto it = begin(factors); it != end(factors); ++it)
-      append(1, *it, flatten_tag);
-  }
+  Product(ExprPtrList factors, Flatten flatten_tag = Flatten::Yes);
 
   /// construct a Product out of zero or more factors (multiplied by 1)
   /// @param rng a range of factors; if rng is a Product, it will be flattened
@@ -198,9 +194,7 @@ class Product : public Expr {
   /// @param factor a factor by which to multiply the product
   /// @param flatten_tag specifies whether (and how) to flatten the argument(s)
   /// @return @c *this
-  Product &append(ExprPtr factor, Flatten flatten_tag = Flatten::Yes) {
-    return this->append(1, factor, flatten_tag);
-  }
+  Product &append(ExprPtr factor, Flatten flatten_tag = Flatten::Yes);
 
   /// (post-)multiplies the product by @c factor
   /// @param factor a factor by which to multiply the product
@@ -268,13 +262,13 @@ class Product : public Expr {
                          flatten_tag);
   }
 
-  const auto &scalar() const { return scalar_; }
+  const scalar_type &scalar() const;
 
   /// @return `Constant::is_zero(this->scalar())`
-  bool is_zero() const override { return Constant::is_zero(this->scalar()); }
+  bool is_zero() const override;
 
-  const auto &factors() const { return factors_; }
-  auto &factors() { return factors_; }
+  const factors_type &factors() const;
+  factors_type &factors();
 
   /// @brief View view of factors that are scalars (anything for which
   ///        Expr::is_scalar() returns true).
@@ -294,10 +288,10 @@ class Product : public Expr {
   /// Factor accessor
   /// @param i factor index
   /// @return ith factor
-  const ExprPtr &factor(size_t i) const { return factors_.at(i); }
+  const ExprPtr &factor(size_t i) const;
 
   /// @return true if the number of factors is zero
-  bool empty() const { return factors_.empty(); }
+  bool empty() const;
 
   /// @brief checks commutativity recursively
   /// @return true if definitely commutative, false definitely not commutative
@@ -312,135 +306,50 @@ class Product : public Expr {
  private:
   /// @return true if commutativity is decidable statically
   /// @sa CProduct::static_commutativity() and NCProduct::static_commutativity()
-  virtual bool static_commutativity() const { return false; }
+  virtual bool static_commutativity() const;
 
  public:
-  std::wstring to_latex() const override { return to_latex(false); }
+  std::wstring to_latex() const override;
 
   /// just like Expr::to_latex() , but can negate before conversion
   /// @param[in] negate if true, scalar will be before conversion
-  std::wstring to_latex(bool negate) const {
-    std::wstring result;
-    result = L"{";
-    if (!scalar().is_zero()) {
-      const auto scal = negate ? -scalar() : scalar();
-      if (!scal.is_identity()) {
-        // replace -1 prefactor by -
-        if (!(negate ? scalar() : -scalar()).is_identity()) {
-          result += io::latex::to_string(scal);
-        } else {
-          result += L"{-}";
-        }
-      }
-      for (const auto &i : factors()) {
-        if (i->is<Product>())
-          result += L"\\bigl(" + i->to_latex() + L"\\bigr)";
-        else
-          result += i->to_latex();
-      }
-    }
-    result += L"}";
-    return result;
-  }
+  std::wstring to_latex(bool negate) const;
 
-  type_id_type type_id() const override { return get_type_id<Product>(); };
+  type_id_type type_id() const override;
 
   /// @return an identical clone of this Product (a deep copy allocated on the
   ///         heap)
   /// @note this does not flatten the product
-  ExprPtr clone() const override { return ex<Product>(this->deep_copy()); }
+  ExprPtr clone() const override;
 
-  Product deep_copy() const {
-    auto cloned_factors =
-        factors() | ranges::views::transform([](const ExprPtr &ptr) {
-          return ptr ? ptr->clone() : nullptr;
-        });
-    Product result(this->scalar(), ExprPtrList{});
-    ranges::for_each(cloned_factors, [&](const auto &cloned_factor) {
-      result.append(1, std::move(cloned_factor), Flatten::No);
-    });
-    return result;
-  }
+  Product deep_copy() const;
 
-  Product &operator*=(const Expr &that) {
-    if (!that.is<Constant>()) {
-      this->append(1, const_cast<Expr &>(that).shared_from_this());
-    } else {
-      scalar_ *= that.as<Constant>().value();
-    }
-    return *this;
-  }
+  Product &operator*=(const Expr &that);
 
-  void add_identical(const Product &other) {
-    SEQUANT_ASSERT(ranges::equal(this->factors(), other.factors()));
-    scalar_ += other.scalar_;
-  }
+  void add_identical(const Product &other);
 
-  void add_identical(const std::shared_ptr<Product> &other) {
-    SEQUANT_ASSERT(ranges::equal(this->factors(), other->factors()));
-    scalar_ += other->scalar_;
-  }
+  void add_identical(const std::shared_ptr<Product> &other);
 
-  void add_identical(const ExprPtr &other) {
-    if (other.is<Product>()) return this->add_identical(other.as<Product>());
+  void add_identical(const ExprPtr &other);
 
-    // only makes sense if this has a single factor
-    SEQUANT_ASSERT(this->factors_.size() == 1 && this->factors_[0] == other);
-    scalar_ += 1;
-  }
+  ExprIterator begin_subexpr() override;
 
-  ExprIterator begin_subexpr() override {
-    if (!factors_.empty()) {
-      reset_hash_value();
-    }
+  ExprIterator end_subexpr() override;
 
-    return ExprIterator{factors_.data()};
-  }
+  ConstExprIterator begin_subexpr() const override;
 
-  ExprIterator end_subexpr() override {
-    return ExprIterator{factors_.data() + factors_.size()};
-  }
-
-  ConstExprIterator begin_subexpr() const override {
-    return ConstExprIterator{factors_.data()};
-  }
-
-  ConstExprIterator end_subexpr() const override {
-    return ConstExprIterator{factors_.data() + factors_.size()};
-  }
+  ConstExprIterator end_subexpr() const override;
 
  private:
   scalar_type scalar_ = {1, 0};
-  container::svector<ExprPtr, 2> factors_{};
+  factors_type factors_{};
 
   /// @return the hash of this object, by hashing only the factors,
   /// not the scalar to make possible rapid finding of Products that only
   /// differ by a factor
   /// @note this ensures that hash of a Product involving a single factor is
   /// identical to the hash of the factor itself.
-  hash_type memoizing_hash() const override {
-    auto compute_hash = [this]() {
-      if (factors().size() == 1)
-        return factors_[0]->hash_value();
-      else {
-        auto deref_factors =
-            factors() |
-            ranges::views::transform(
-                [](const ExprPtr &ptr) -> const Expr & { return *ptr; });
-        auto value = hash::range(ranges::begin(deref_factors),
-                                 ranges::end(deref_factors));
-        return value;
-      }
-    };
-
-    if (!hash_value_) {
-      hash_value_ = compute_hash();
-    } else {
-      SEQUANT_ASSERT(*hash_value_ == compute_hash());
-    }
-
-    return *hash_value_;
-  }
+  hash_type memoizing_hash() const override;
 
   ExprPtr canonicalize_impl(CanonicalizeOptions);
 
@@ -454,21 +363,7 @@ class Product : public Expr {
               CanonicalizationMethod::Rapid)) override;
 
  private:
-  bool static_equal(const Expr &that) const override {
-    const auto &that_cast = static_cast<const Product &>(that);
-    if (scalar() == that_cast.scalar() &&
-        factors().size() == that_cast.factors().size()) {
-      if (this->empty()) return true;
-      // compare hash values first
-      if (this->hash_value() ==
-          that.hash_value())  // hash values agree -> do full comparison
-        return std::equal(begin_subexpr(), end_subexpr(), that.begin_subexpr(),
-                          expr_ptr_comparer);
-      else
-        return false;
-    } else
-      return false;
-  }
+  bool static_equal(const Expr &that) const override;
 };  // class Product
 
 class CProduct : public Product {
